✨ feat: Update layout in README.md and index.html
This commit updates the layout in both README.md and index.html. In README.md, a grid of images has been added to provide an overview of the rooms and their equipment. The grid now consists of four columns instead of two. In index.html, the column count for the grid has also been changed to four.