✨ feat: Add new drawing option to command palette and update last open files
This commit adds a new feature to the command palette by including the option to "Open command palette" for the "obsidian-excalidraw-plugin". Additionally, it updates the list of last open files in the workspace.json file to include "_media/rooms/zimmer-two-wide.jpg".

---

chore: Update README.md with room images

The README.md file has been updated to include images of different rooms. The added images are:
- Küche
- Badezimmer
- Zimmer - Ein-Bett-Konfiguration
- Zimmer - Zwei-Bett-Konfiguration