✨ feat: Add information about Smart Home and temperature/humidity sensors
Added a section in the README.md file to provide information about the Smart Home features in the apartment. Mentioned the presence of Alexa voice assistants that can be addressed as "Computer" and used for playing Deutschlandfunk Nova or as a timer/alarm. Also mentioned that these devices can be unplugged if not needed. Noted that there are no voice assistants in the guest's room, only a router and a small computer in a basket on the shelf.

Additionally, included details about the automatic temperature and humidity sensors used to prevent mold formation. Described them as small white sensors usually placed on the door frames.

Modified index.html to adjust the styling of the grid figcaption element, reducing the minimum height to 2.8em.

Files modified:
- README.md
- index.html