✨ feat: Add information about temperature and humidity sensors
The commit adds information about the automatic measurement of temperature and humidity using sensors. The purpose of these measurements is to prevent the formation of mold. The sensors are small and white and usually placed on the door frames. Additionally, a portion of the measurement data can be read in the hallway mirror.