✨ feat: Add translation workflows and README files for Czech and Polish languages
This commit adds new workflow steps to translate the README file into Czech and Polish languages. The translation is done using the dephraiim/translate-readme@main action.

Changes made:
- Created a new step "Adding README - Czech" that uses the translate-readme action with LANG set to cz.
- Created a new step "Adding README - Polski" that uses the translate-readme action with LANG set to pl.

In addition, the _coverpage.md files for both the root directory and the en directory were updated to include links to the translated README files for Czech and Polish languages.