✨ feat: Add information about Alexa Sprachassistenten
The README.md file has been updated to include information about the Alexa Sprachassistenten in the apartment. The Alexa devices can be addressed by the name "Computer" and can play Deutschlandfunk Nova. They are available in the bathroom and kitchen, and they understand both German and English. It is also mentioned that if they are not needed, they can be disconnected from power. Additionally, it is clarified that there are no Sprachassistenten in the tenant's room, only a router and a small computer located in a basket on the shelf.