🔧 chore: Add translation workflows and READMEs for Urdu and Vietnamese
This commit adds the necessary steps to the translation workflow to include the translation of the README into Urdu and Vietnamese. The translate-readme action is used with the respective language codes ("ur" for Urdu and "vi" for Vietnamese) to generate the translated README files. Additionally, the _coverpage.md file is updated to include the language code and name for Urdu in the list of supported languages.